# enterprise/sdk/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: HeadyOS Platform API
  description: |
    Official API for HeadyOS Platform and HeadyMe AI (headyme.com).
    
    HeadySystems Inc. (DBA Heady™) — https://headyme.com
    
    All numeric parameters derive from φ = 1.618033988749895 and Fibonacci sequences:
    - Rate limits: Fibonacci burst rates (fib(12)=144 Pilot, fib(14)=377 Pro, fib(16)=987 Enterprise req/min)
    - Timeout defaults: 1000ms × φ^n
    - Retry backoff: 1000ms × φ^n
    
    ## Authentication
    Use Bearer JWT or API Key authentication via `Authorization: Bearer <token>` header.
    
    ## Rate Limits
    Rate limit headers returned on every response:
    - `X-RateLimit-Limit`: Request limit per window
    - `X-RateLimit-Remaining`: Requests remaining
    - `X-RateLimit-Reset`: Reset timestamp (Unix)
    - `X-RateLimit-Window`: Window duration in seconds
  version: 1.0.0
  contact:
    name: HeadySystems SDK Support
    email: sdk@headyme.com
    url: https://docs.headyme.com
  license:
    name: MIT
    url: https://github.com/heady-ai/sdk-javascript/blob/main/LICENSE
  x-logo:
    url: https://headyme.com/logo.png
    altText: HeadyOS
servers:
  - url: https://api.headyme.com/v1
    description: Production
  - url: https://api-staging.headyme.com/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Local development
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# ---------------------------------------------------------------------------
# Tags
# ---------------------------------------------------------------------------
...
paths:
...
  /brain/chat:
    post:
      tags: [Brain]
      operationId: brainChat
      summary: Chat with AI brain
      description: |
        Send messages to the Heady™OS AI brain (GPT-4o, Claude, Llama, etc.)
        with optional MCP tool access and vector memory context.
        
        Default temperature: 1/φ ≈ 0.618 (HIGH CSL threshold)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
            example:
              messages:
                - role: "user"
                  content: "What is multi-agent orchestration?"
              temperature: 0.618
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
  /brain/analyze:
    post:
      tags: [Brain]
      operationId: brainAnalyze
      summary: Analyze text with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeResponse"

  # Agents
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT Bearer token from /auth/token or /auth/token/refresh"
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API Key: 'Bearer hdy_your_api_key'"
...
  responses:
...
    ValidationError:
      description: Input validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
        X-RateLimit-Reset:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RateLimitErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
        requestId:
          type: string
        retryable:
          type: boolean
    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            issues:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
                  code:
                    type: string
    RateLimitErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            retryAfterMs:
              type: integer
            limit:
              type: integer
            remaining:
              type: integer
            resetAt:
              type: string
              format: date-time
...
    Message:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant, system, function]
        content:
          type: string
        name:
          type: string
        functionCall:
          type: object
          properties:
            name:
              type: string
            arguments:
              type: string
    ChatRequest:
      type: object
      required: [messages]
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        model:
          type: string
        temperature:
          type: number
          minimum: 0
          maximum: 1
          default: 0.618
          description: "Default: 1/φ ≈ 0.618 (HIGH CSL threshold)"
        maxTokens:
          type: integer
        stream:
          type: boolean
          default: false
        systemPrompt:
          type: string
        agentId:
          type: string
        memoryNamespace:
          type: string
    ChatResponse:
      type: object
      properties:
        id:
          type: string
        model:
          type: string
        message:
          $ref: "#/components/schemas/Message"
        usage:
          type: object
          properties:
            promptTokens:
              type: integer
            completionTokens:
              type: integer
            totalTokens:
              type: integer
        finishReason:
          type: string
        latencyMs:
          type: integer
        createdAt:
          type: string
          format: date-time
    AnalyzeRequest:
      type: object
      required: [input]
      properties:
        input:
          type: string
        analysisType:
          type: string
          default: reasoning
        outputFormat:
          type: string
          enum: [json, text, markdown]
        confidence:
          type: boolean
        model:
          type: string
    AnalyzeResponse:
      type: object
      properties:
        id:
          type: string
        analysisType:
          type: string
        result:
          description: Analysis result (format depends on outputFormat)
        confidence:
          type: number
          minimum: 0
          maximum: 1
        model:
          type: string
        latencyMs:
          type: integer
        createdAt:
          type: string
          format: date-time

    # Agents
...